+++ src/places_crawler.js
@@ -6,11 +6,12 @@ const typedefs = require('./typedefs'); // eslint-disable-line no-unused-vars
 const { enqueueAllPlaceDetails } = require('./enqueue_places');
 const { handlePlaceDetail } = require('./detail_page_handle');
 const {
-    waitAndHandleConsentScreen, waiter,
+    waitAndHandleConsentScreen, waiter, blockRequestsForOptimization,
 } = require('./utils/misc-utils');
+const { LABELS } = require('./consts');
 
 const { log } = Apify.utils;
-const { injectJQuery, blockRequests } = Apify.utils.puppeteer;
+const { injectJQuery } = Apify.utils.puppeteer;
 
 /**
  * @param {{
@@ -25,8 +26,6 @@ const handlePageFunctionExtended = async ({ pageContext, scrapingOptions, helper
 
     const { label, searchString } = /** @type {{ label: string, searchString: string }} */ (request.userData);
 
-    const logLabel = label === 'startUrl' ? 'SEARCH' : 'PLACE';
-
     // TODO: Figure out how to remove the timeout and still handle consent screen
     // Handle consent screen, this wait is ok because we wait for selector later anyway
     await page.waitForTimeout(5000);
@@ -34,7 +33,10 @@ const handlePageFunctionExtended = async ({ pageContext, scrapingOptions, helper
     // @ts-ignore I'm not sure how we could fix the types here
     if (request.userData.waitingForConsent !== undefined) {
         // @ts-ignore  I'm not sure how we could fix the types here
-        await waiter(() => request.userData.waitingForConsent === false);
+        await waiter(
+            () => request.userData.waitingForConsent === false,
+            { timeout: 70000, timeoutErrorMeesage: 'Waiting for cookie consent timeouted, reloading page again' }
+        );
     }
 
     // Inject JQuery crashes on consent screen
@@ -44,14 +46,14 @@ const handlePageFunctionExtended = async ({ pageContext, scrapingOptions, helper
     try {
         // Check if Google shows captcha
         if (await page.$('form#captcha-form')) {
-            throw `[${logLabel}]: Got CAPTCHA on page, retrying --- ${searchString || ''} ${request.url}`;
+            throw `[${label}]: Got CAPTCHA on page, retrying --- ${searchString || ''} ${request.url}`;
         }
-        if (label === 'startUrl') {
+        if (label === LABELS.SEARCH) {
             if (!maxCrawledPlacesTracker.canEnqueueMore(searchString || request.url)) {
                 // No need to log anything here as it was already logged for this search
                 return;
             }
-            log.info(`[${logLabel}]: Start enqueuing places details for search --- ${searchString || ''} ${request.url}`);
+            log.info(`[${label}]: Start enqueuing places details for search --- ${searchString || ''} ${request.url}`);
             await errorSnapshotter.tryWithSnapshot(
                 page,
                 async () => enqueueAllPlaceDetails({
@@ -65,11 +67,11 @@ const handlePageFunctionExtended = async ({ pageContext, scrapingOptions, helper
                 }),
             );
 
-            log.info(`[${logLabel}]: Enqueuing places finished for --- ${searchString || ''} ${request.url}`);
+            log.info(`[${label}]: Enqueuing places finished for --- ${searchString || ''} ${request.url}`);
             stats.maps();
-        } else {
+        } else if (label === LABELS.PLACE) {
             // Get data for place and save it to dataset
-            log.info(`[${logLabel}]: Extracting details from place url ${page.url()}`);
+            log.info(`[${label}]: Extracting details from place url ${page.url()}`);
 
             await handlePlaceDetail({
                 page,
@@ -83,6 +85,9 @@ const handlePageFunctionExtended = async ({ pageContext, scrapingOptions, helper
                 maxCrawledPlacesTracker,
                 crawler,
             });
+        } else {
+            // This is developer error, should never happen
+            throw new Error(`Unkown label "${label}" provided in the Request to the Crawler`);
         }
         stats.ok();
     } catch (err) {
@@ -120,18 +125,8 @@ module.exports.setUpCrawler = ({ crawlerOptions, scrapingOptions, helperClasses 
             
             const mapUrl = new URL(request.url);
 
-            // Never block images for allPlacesNoSearch to keep pins visible
-            if (!allPlacesNoSearchAction && !maxImages) {
-                // https://lh5.googleusercontent.com/p/AF1QipMInapT8CB8U-QFRfRceZtzxbX5QRw0NJ08Fc7t=w408-h272-k-no
-                // We need map working for search scrolling
-                const urlPatterns = request.userData.label === 'PLACE'
-                    ? ['/maps/vt/', '/earth/BulkMetadata/', 'googleusercontent.com']
-                    : [];
-                await blockRequests(page, {
-                    urlPatterns,
-                });
-            }
-            
+            await blockRequestsForOptimization(page, request.userData.label, maxImages, allPlacesNoSearchAction);
+
             if (language) {
                 mapUrl.searchParams.set('hl', language);
             }